# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Hexagonal blog platform REST API. Envelope responses use { ok, code, data, meta, error }.",
        "title": "Blog API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/turahe/blog-api"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/media": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List media assets",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "disk filter",
                        "name": "disk",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Presign media upload",
                "parameters": [
                    {
                        "description": "upload metadata",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.PresignMedia"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/media/{param1}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete media asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "media UUID",
                        "name": "param1",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/media/{param1}/complete": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Complete media upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "media UUID",
                        "name": "param1",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/media/{param1}/tags": {
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Patch media tags",
                "parameters": [
                    {
                        "type": "string",
                        "description": "media UUID",
                        "name": "param1",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "tags",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.PatchMediaTags"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/media/{param1}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Get ready media asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "media UUID",
                        "name": "param1",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.PatchMediaTags": {
            "type": "object",
            "required": [
                "tags"
            ],
            "properties": {
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "requests.PresignMedia": {
            "type": "object",
            "required": [
                "content_type",
                "original_filename",
                "size_bytes"
            ],
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "original_filename": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "size_bytes": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "responses.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "error": {
                    "$ref": "#/definitions/responses.ErrorBody"
                },
                "links": {
                    "$ref": "#/definitions/responses.PageLinks"
                },
                "meta": {
                    "type": "object"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "responses.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.PageLinks": {
            "type": "object",
            "properties": {
                "first": {
                    "type": "string"
                },
                "last": {
                    "type": "string"
                },
                "next": {
                    "type": "string"
                },
                "prev": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}